flaskk/requirements.txt
flask
google-generativeai
dotenv-python